requirements.txt
PyMuPDF==1.25.4
pdf2image==1.17.0
pytesseract==0.3.13
flask==3.0.2
werkzeug==3.0.1
pdfplumber==0.10.3
requests==2.31.0
youtube-transcript-api==0.6.2
openai-whisper==20231117
moviepy==1.0.3 